+++ hawtio-system/src/test/java/io/hawt/web/ProxyDetailsTest.java
@@ -1,7 +1,8 @@
 package io.hawt.web;
 
 import java.util.Arrays;
-import java.util.List;
+import java.util.HashSet;
+import java.util.Set;
 
 import org.junit.Test;
 
@@ -29,7 +30,7 @@ public class ProxyDetailsTest {
         assertEquals("getPort()", 8181, details.getPort());
         assertEquals("getProxyPath()", "/jolokia/", details.getProxyPath());
         assertEquals("getScheme()", "http", details.getScheme());
-        assertEquals("getStringProxyURL()", "http://localhost:8181/jolokia/", details.getFullProxyUrl());
+        assertEquals("getFullProxyUrl()", "http://localhost:8181/jolokia/", details.getFullProxyUrl());
     }
 
     @Test
@@ -46,7 +47,7 @@ public class ProxyDetailsTest {
         assertEquals("getPort()", 80, details.getPort());
         assertEquals("getProxyPath()", "/jolokia/", details.getProxyPath());
         assertEquals("getScheme()", "http", details.getScheme());
-        assertEquals("getStringProxyURL()", "http://localhost/jolokia/", details.getFullProxyUrl());
+        assertEquals("getFullProxyUrl()", "http://localhost/jolokia/", details.getFullProxyUrl());
     }
 
     @Test
@@ -63,7 +64,7 @@ public class ProxyDetailsTest {
         assertEquals("getPort()", 80, details.getPort());
         assertEquals("getProxyPath()", "/jolokia/", details.getProxyPath());
         assertEquals("getScheme()", "http", details.getScheme());
-        assertEquals("getStringProxyURL()", "http://localhost/jolokia/", details.getFullProxyUrl());
+        assertEquals("getFullProxyUrl()", "http://localhost/jolokia/", details.getFullProxyUrl());
     }
 
     @Test
@@ -79,7 +80,7 @@ public class ProxyDetailsTest {
         assertEquals("getPort()", 90, details.getPort());
         assertEquals("getProxyPath()", "/jolokia/", details.getProxyPath());
         assertEquals("getScheme()", "http", details.getScheme());
-        assertEquals("getStringProxyURL()", "http://localhost:90/jolokia/", details.getFullProxyUrl());
+        assertEquals("getFullProxyUrl()", "http://localhost:90/jolokia/", details.getFullProxyUrl());
     }
 
     @Test
@@ -95,7 +96,7 @@ public class ProxyDetailsTest {
         assertEquals("getPort()", 80, details.getPort());
         assertEquals("getProxyPath()", "/cxf/crm/customerservice/customers/123", details.getProxyPath());
         assertEquals("getScheme()", "http", details.getScheme());
-        assertEquals("getStringProxyURL()", "http://somerest-davsclaus2.rhcloud.com/cxf/crm/customerservice/customers/123", details.getFullProxyUrl());
+        assertEquals("getFullProxyUrl()", "http://somerest-davsclaus2.rhcloud.com/cxf/crm/customerservice/customers/123", details.getFullProxyUrl());
     }
 
     @Test
@@ -112,7 +113,7 @@ public class ProxyDetailsTest {
         assertEquals("getPort()", 443, details.getPort());
         assertEquals("getProxyPath()", "/myApp/jolokia/", details.getProxyPath());
         assertEquals("getScheme()", "https", details.getScheme());
-        assertEquals("getStringProxyURL()", "https://www.myhost.com/myApp/jolokia/", details.getFullProxyUrl());
+        assertEquals("getFullProxyUrl()", "https://www.myhost.com/myApp/jolokia/", details.getFullProxyUrl());
     }
 
     @Test
@@ -129,7 +130,7 @@ public class ProxyDetailsTest {
         assertEquals("getPort()", 443, details.getPort());
         assertEquals("getProxyPath()", "/myApp/jolokia/", details.getProxyPath());
         assertEquals("getScheme()", "https", details.getScheme());
-        assertEquals("getStringProxyURL()", "https://www.myhost.com/myApp/jolokia/", details.getFullProxyUrl());
+        assertEquals("getFullProxyUrl()", "https://www.myhost.com/myApp/jolokia/", details.getFullProxyUrl());
     }
 
     @Test
@@ -146,7 +147,7 @@ public class ProxyDetailsTest {
         assertEquals("getPort()", 443, details.getPort());
         assertEquals("getProxyPath()", "/myApp/jolokia/", details.getProxyPath());
         assertEquals("getScheme()", "https", details.getScheme());
-        assertEquals("getStringProxyURL()", "https://www.myhost.com/myApp/jolokia/", details.getFullProxyUrl());
+        assertEquals("getFullProxyUrl()", "https://www.myhost.com/myApp/jolokia/", details.getFullProxyUrl());
     }
 
     @Test
@@ -165,7 +166,7 @@ public class ProxyDetailsTest {
         assertEquals("getPort()", 443, details.getPort());
         assertEquals("getProxyPath()", "/myApp/jolokia/", details.getProxyPath());
         assertEquals("getScheme()", "https", details.getScheme());
-        assertEquals("getStringProxyURL()", "https://www.myhost.com/myApp/jolokia/?foo=bar", details.getFullProxyUrl());
+        assertEquals("getFullProxyUrl()", "https://www.myhost.com/myApp/jolokia/?foo=bar", details.getFullProxyUrl());
     }
 
     @Test
@@ -183,7 +184,7 @@ public class ProxyDetailsTest {
         assertEquals("getPort()", 443, details.getPort());
         assertEquals("getProxyPath()", "/myApp/jolokia/", details.getProxyPath());
         assertEquals("getScheme()", "https", details.getScheme());
-        assertEquals("getStringProxyURL()", "https://www.myhost.com/myApp/jolokia/", details.getFullProxyUrl());
+        assertEquals("getFullProxyUrl()", "https://www.myhost.com/myApp/jolokia/", details.getFullProxyUrl());
     }
 
     @Test
@@ -201,7 +202,7 @@ public class ProxyDetailsTest {
         assertEquals("getPort()", 443, details.getPort());
         assertEquals("getProxyPath()", "/myApp/jolokia/", details.getProxyPath());
         assertEquals("getScheme()", "https", details.getScheme());
-        assertEquals("getStringProxyURL()", "https://www.myhost.com/myApp/jolokia/", details.getFullProxyUrl());
+        assertEquals("getFullProxyUrl()", "https://www.myhost.com/myApp/jolokia/", details.getFullProxyUrl());
     }
 
     @Test
@@ -219,7 +220,7 @@ public class ProxyDetailsTest {
         assertEquals("getPort()", 443, details.getPort());
         assertEquals("getProxyPath()", "/myApp/jolokia/", details.getProxyPath());
         assertEquals("getScheme()", "https", details.getScheme());
-        assertEquals("getStringProxyURL()", "https://www.myhost.com/myApp/jolokia/?search=1234&page=4", details.getFullProxyUrl());
+        assertEquals("getFullProxyUrl()", "https://www.myhost.com/myApp/jolokia/?search=1234&page=4", details.getFullProxyUrl());
     }
 
     @Test
@@ -237,7 +238,7 @@ public class ProxyDetailsTest {
         assertEquals("getPort()", 80, details.getPort());
         assertEquals("getProxyPath()", "/jolokia//exec/org.apache.camel:context=camel-1,type=context,name=%22camel-1%22/canSendToEndpoint(java.lang.String)/activemq:!/!/queue:newOrder", details.getProxyPath());
         assertEquals("getScheme()", "http", details.getScheme());
-        assertEquals("getStringProxyURL()", "http://127.0.0.1:54155/jolokia//exec/org.apache.camel:context=camel-1,type=context,name=%22camel-1%22/canSendToEndpoint(java.lang.String)/activemq:!/!/queue:newOrder?maxDepth=7&maxCollectionSize=5000&ignoreErrors=true&canonicalNaming=false", details.getFullProxyUrl());
+        assertEquals("getFullProxyUrl()", "http://127.0.0.1:54155/jolokia//exec/org.apache.camel:context=camel-1,type=context,name=%22camel-1%22/canSendToEndpoint(java.lang.String)/activemq:!/!/queue:newOrder?maxDepth=7&maxCollectionSize=5000&ignoreErrors=true&canonicalNaming=false", details.getFullProxyUrl());
     }
 
     @Test
@@ -249,7 +250,7 @@ public class ProxyDetailsTest {
             .thenReturn("/www.myhost.com/jolokia/")
             .thenReturn("/www.banned.com/jolokia/");
 
-        List<String> whitelist = Arrays.asList("localhost", "www.myhost.com");
+        Set<String> whitelist = new HashSet<>(Arrays.asList("localhost", "www.myhost.com"));
         ProxyDetails details1 = new ProxyDetails(mockReq);
         ProxyDetails details2 = new ProxyDetails(mockReq);
         ProxyDetails details3 = new ProxyDetails(mockReq);
@@ -268,7 +269,7 @@ public class ProxyDetailsTest {
             .thenReturn("/www.myhost.com/jolokia/")
             .thenReturn("/www.banned.com/jolokia/");
 
-        List<String> whitelist = Arrays.asList("*");
+        Set<String> whitelist = new HashSet<>(Arrays.asList("*"));
         ProxyDetails details1 = new ProxyDetails(mockReq);
         ProxyDetails details2 = new ProxyDetails(mockReq);
         ProxyDetails details3 = new ProxyDetails(mockReq);